spu 保存，propertyValue合并到productProperty中

+++ yudao-module-mall/yudao-module-product-biz/src/main/java/cn/iocoder/yudao/module/product/service/property/ProductPropertyService.java
@@ -81,5 +81,5 @@ public interface ProductPropertyService {
      * @param propertyIds 属性名id集合
      * @return
      */
-    List<ProductPropertyRespVO> selectByIds(List<Integer> propertyIds);
+    List<ProductPropertyRespVO> selectByIds(List<Long> propertyIds);
 }
